fix(auth): 修复菜单权限服务中的计数类型转换问题
- 将MenuServiceImpl中的menuCount从Integer改为Long类型
- 使用Math.toIntExact方法安全地将Long类型的menuCount转换为int创建ArrayList
- 在PrivilegeServiceImpl中将count变量类型从Integer改为Long
- 在RoleServiceImpl中将两个count变量类型都从Integer改为Long以保持一致性

+++ tj-auth/tj-auth-service/src/main/java/com/tianji/auth/service/impl/MenuServiceImpl.java
@@ -117,12 +117,12 @@ public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu> implements IM
             throw new CommonException(ROLE_NOT_FOUND);
         }
         // 2.判断菜单是否存在
-        Integer menuCount = lambdaQuery().in(Menu::getId, menuIds).count();
+        Long menuCount = lambdaQuery().in(Menu::getId, menuIds).count();
         if (menuCount != menuIds.size()) {
             throw new CommonException(MENU_NOT_FOUND);
         }
         // 3.绑定关系
-        List<RoleMenu> roleMenus = new ArrayList<>(menuCount);
+        List<RoleMenu> roleMenus = new ArrayList<>(Math.toIntExact(menuCount));
         for (Long menuId : menuIds) {
             roleMenus.add(new RoleMenu(roleId, menuId));
         }
